// client/src/api-clien/models/CreateUserDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type CreateUserDto = {
    /**
     * User login
     */
    login: string;
    /**
     * User password
     */
    password: string;
    /**
     * User nickname
     */
    nickname: string;
    /**
     * User email
     */
    email: string;
};